Added feature to control the display of the Show/Hide button on the UserCode form
This feature allows the administrator to control whether the Show/Hide button on the UserCode
form is displayed using a checkbox.  The state of this checkbox is stored in the database
under show_password.

Note: show_password in the database is associated with the Show/Hide code
toggle button on the UserCode form and not the Show/Hide password toggle button on the
admin password page

When the Show/Hide button is hidden, the Show/Hide state will be set to hide the user
code and remain in this state until the Show/Hide button is displayed again.