Resolved intermittent fingerprint enabling behavior
There was an issue with editing codes that resulted in fingerprint
to become enabled intermittently.  It turns out that when reading
from the database the conversion from int to bool was unpredictable.
Even though the database showed 0 and the print after getting the
field from the database showed 0, the resulting conversion became 1.
Added an explicit int to bool conversion where the bool because true
if and only if the int value was 1.  Anything else is false.  This
seems to have resolved the issue.

Changed the identify fingerprint button on the usercode page to
be 'not checkable'.  Before the button would remain pushed in after
enter a fingerprint which seem odd and unintuitive.  Now the button
clicks, launches the fingerprint dialog and returns to normal unclicked
state.