Finished multi-lock feature and testing for code 1
A large number of changes were necessary to multi-lock feature.
The most invasive was the change from int for lock number to string
for comma-separated list of lock numbers.  Also, new widgets were
add: lock-cabinet, select locks, keyboard.  Some of this functionality
existed previously but no in a reusable form.  Now there are widgets
that can be placed as needed.