Completed fingerprint integration with multilock
Had to make changes to wrt to the fingerprint dialog and interactions
with system controller to support multi-lock.  Also, moved fingerprint
verify dialog from keycodebox main to system controller.  There was
no interaction with keycoebox main so it wasn't the correct location,
but also, system controller is the decision point for multi lock and
it needed to control the visibility of fingerprint verify.