Modified lock history features to support multiple locks
With the addition of multiple locks, it was decided to keep
history logging the same, i.e., one entry per lock.  The existing
architecture oddly and unnecessarily propagates data up and down
the stack via signal/slot pairs as opposed to signal/signal notification.
The result is the loss of what locks are actually opened.  A small
modification was made to force writing the history as each lock is
opened.

The underlying storage of the lock history data in lock history set
was changed from a mapping list (which did not preserve ordering
from the SQL query) to a QVector -- which also required changes to
the interator.