Added support for display power down
In cases where there is limited power, i.e., battery, solar, etc,
the display takes a lot of power.  The Pi firmware supports commands
that allow the hdmi power to be turned on/off which effectively
blanks the display.  Tests will be needed to measure and confirm
how much power is saved.

The default implementation uses a 30 second timer to power down the
display.  When the app returns to the logo screen, the timer starts.
If no touch event occurs, the display will power down in 30 seconds.
If a touch event is detected, the timer is stopped and normal operation
resumes.  Upon displaying the logo screen again, the timer is restarted.
Also, when powered down, two touches are required to enter a code.  The
first touch powers on the display to show the logo screen.  The second
touch operates as normal.
It is not possible, as far as I know, to read the power state, so the
application assumes on startup that the display should be on.  This is
reasonable because the Pi boots into the application and the display
power state is not persistent and is powered on by default.
Some enhancements are possible: select whether power down should be
active and specify the timeout.