fixed crash when all locks selected
The lock cabinet widget was using the lock value (1-based)
instead of the lock index value (0-based).  This appeared work
correctly until lock 32 was selected.  In this case, the cabinet
index calculation became 1 instead of 0 and caused the crash.

Additionally, found a problem with editing a code that had lock 1
selected, and changing to select all locks.  In this case, all the
locks except 1 were selected.  This occurred because of a condition
checking if there was a change.  That check was unnecessary and
was removed.