Fixed issue with indexing in lock cabinet widget
Changed vector of locks from QVector<int> to QVector<QString>.
With int as a type, there were a lot of unnecessary conversions between string
and int.  With this change, the only conversion is needed when determining
the cabinet and lock indecies.  Added some helper methods for added/removing
selected locks.  There was a bug in the selected locks management that wasn't
removing the lock after it was opened.  This may have contributed to the
multiple lock open commands sent that Buzz was seeing.