Made numerous changes to report to fix old issues and make usage clearer
The report page was very messy and unstructured.  Also, it did not enforce
control of the ui based on feature selection/enablement.

1. Changed auto report - feature is not enabled/disabled via a group checkbox.
Removed the 'Never' option from Frequency.  The enable/disable for auto report
is controlled via logic for report storage, i.e., auto report is only enabled if
there is either or both email report or report to file enabled.
2. Changed immediate report - feature enable/disable controlled similar to auto
report.
3. Introduced new concept, Report Storage.  Report storage covers 'send via email'
and 'save to file' functionalities.  'Send via email' is essentially the same.
'Report to file' is sigificantly different.  There are two facets of 'Report to file':
save files to default location and save files to USB drive.  Save to default location
is the only possible behavior if there is no USB drive inserted.  Additionally,
in order to keep from overflowing the Pi storage, the default location is monitored
and reports older than a specific time frame will be deleted automatically.  The default
deletion period is 1 month, but the user can selected shorter periods, 1 day, 1 week, etc.
As reports are created they are displayed in the 'Available Reports' list.  When the list
is populated, reports can be selected and either deleted and/or downloaded.  A USB drive
is needed to download reports.  When saving to a USB drive, the 'Store to USB Drive' will
be populated if a USB drive is inserted.  The user can enable 'Store to USB Drive' and
select the drive.  The storage directory will be '/media/pi/<USB Drive Name>.  There is
no management of the USB drive storage, e.g., no auto deletion.  If there are existing
reports when 'Save Report to File' is disabled, behavior depends on the storage mode.
If reports are stored to the default location, then the reports will be deleted.  The
user is notified and instructed how to save reports, if they want to.  If reports are
stored to USB drive then they will no be deleted.