Added new module to run system commands
Through out the code there is a disjointed set of functions that send
a system and command read back the results.  They uses 'system' and
a poll loop to read back the results.  I decided that this should be
done using Qt.  The kcbsystem h/cpp files contain the basics for
sending system commands and reading back from stdout and/or stderr.
Over time I will replace the other approach with this one and expand
the kcbsystem module so that all system commands are in one place.