Added feature to disable the display when connected via VNC
Two customers raised a concern that the display is shown while
they access the system remotely via VNC.  Requested that the box
display be disabled when accessing administrative features via
VNC.

Solution: there are several VNC product available that support
disabling the display but all require a license.  Recently, we
added support to power off the display in order to reduce power
consumption.  The feature turns off power to the HDMI port as if
the cable were removed.  The power feature only affects the display
it does not affect VNC.  Using this same approach, I added selective
calls to power off/on the HDMI port while entering/exiting admin
functionality.