Completed fix to display more than 18 characters and support updating the app binary
Discovered in version 1.10.1 for Chevin, the user code display was limited to 18 characters.  This limitation was found when
using an HID with 19 digits resulting in the wrong code being displayed.

It would have nice if we could have just sent the binary and updated it, but that feature was broken.  Made changes to fix
application binary update.  Included both application code change and kcb-config script changes.
# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
# On branch hotfix-1.10.1
# Changes to be committed:
#	modified:   frmadmininfo.cpp
#	modified:   frmadmininfo.ui
#	modified:   frmusercode.ui
#	modified:   kcbsystem.cpp
#	modified:   kcbsystem.h
#	modified:   version.h
#